+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Generate OLM bundle ([#645])
+- Cluster status conditions ([#xxx])
 
 ### Changed
 
@@ -21,6 +22,7 @@ All notable changes to this project will be documented in this file.
 [#649]: https://github.com/stackabletech/zookeeper-operator/pull/649
 [#651]: https://github.com/stackabletech/zookeeper-operator/pull/651
 [#653]: https://github.com/stackabletech/zookeeper-operator/pull/653
+[#xxx]: https://github.com/stackabletech/zookeeper-operator/pull/xx
 
 ## [23.1.0] - 2023-01-23